renamed wavelet_new to wavelet, removed old version, verified that all except pattern_compressor runs well. Need to update the latter
